Extract terminal I/O into guessthedis/terminal.py
Move Ansi, printc, ignore_sigint, NavigationRequested, read_line, and
pick_challenge into a dedicated terminal module. __main__.py now only
contains disassembly quiz logic and the CLI entrypoint.
